separated utils to avoid circular imports
