+++ examples/armbench-segmentation/armbench_segmentation/utils/general_utils.py
@@ -3,17 +3,14 @@ from typing import Union, Optional, List
 
 import numpy as np
 import tensorflow as tf
-from code_loader.contract.datasetclasses import PreprocessResponse
 from code_loader.contract.responsedataclasses import BoundingBox
 from code_loader.helpers.detection.utils import xyxy_to_xywh_format, xywh_to_xyxy_format
 from code_loader.helpers.detection.yolo.utils import reshape_output_list
 from numpy._typing import NDArray
 
 from armbench_segmentation import CACHE_DICTS
-from armbench_segmentation.preprocessing import BATCH_SIZE, CATEGORIES, MODEL_FORMAT, BACKGROUND_LABEL, IMAGE_SIZE, \
-    DEFAULT_BOXES
-from armbench_segmentation.visualizers import multiple_mask_pred, multiple_mask_gt
-from armbench_segmentation.yolo_helpers.yolo_utils import DECODER
+from armbench_segmentation.preprocessing import BATCH_SIZE, CATEGORIES, MODEL_FORMAT, BACKGROUND_LABEL, IMAGE_SIZE
+from armbench_segmentation.yolo_helpers.yolo_utils import DECODER, DEFAULT_BOXES
 
 
 def polygon_to_bbox(vertices):
@@ -47,53 +44,7 @@ def polygon_to_bbox(vertices):
     return bbox
 
 
-def ioa_mask(mask_containing, mask_contained):
-    """
-    Calculates the Intersection over Area (IOA) between two binary masks.
 
-    Args:
-        mask_containing (ndarray or Tensor): Binary mask representing the containing object.
-        mask_contained (ndarray or Tensor): Binary mask representing the contained object.
-
-    Returns:
-        float: The IOA (Intersection over Area) value between the two masks.
-
-    Note:
-        - The input masks should have compatible shapes.
-        - The function performs a bitwise AND operation between the 'mask_containing' and 'mask_contained' masks to obtain
-          the intersection mask.
-        - It calculates the number of True values in the intersection mask to determine the intersection area.
-        - The area of the contained object is computed as the number of True values in the 'mask_contained' mask.
-        - If the area of the contained object is 0, the IOA is defined as 0.
-        - The IOA value is calculated as the ratio of the intersection area to the maximum of the area of the contained
-          object or 1.
-    """
-
-    intersection_mask = mask_containing & mask_contained
-    intersection = len(intersection_mask[intersection_mask])
-    area = len(mask_contained[mask_contained])
-    return intersection / max(area, 1)
-
-
-def get_ioa_array(image, y_pred_bb, y_pred_mask, bb_gt, mask_gt, containing='pred'):
-    hash_str = str(image) + str(y_pred_bb) + str(y_pred_mask) + str(bb_gt) + str(mask_gt) + str(containing)
-    res = CACHE_DICTS['get_ioa_array'].get(hash_str)
-    if res is not None:
-        return res
-    prediction_masks = multiple_mask_pred(image, y_pred_bb, y_pred_mask)
-    gt_masks = multiple_mask_gt(image, bb_gt, mask_gt)
-    ioas = np.zeros((len(prediction_masks), len(gt_masks)))
-    for i, pred_mask in enumerate(prediction_masks):
-        for j, gt_mask in enumerate(gt_masks):
-            if containing == 'pred':
-                ioas[i, j] = ioa_mask(pred_mask, gt_mask)
-            else:
-                ioas[i, j] = ioa_mask(gt_mask, pred_mask)
-    if len(CACHE_DICTS['get_ioa_array'].keys()) > 2 * BATCH_SIZE:
-        CACHE_DICTS['get_ioa_array'] = {hash_str: ioas}
-    else:
-        CACHE_DICTS['get_ioa_array'][hash_str] = ioas
-    return ioas
 
 
 def calculate_iou_all_pairs(bboxes: np.ndarray, image_size: int) -> np.ndarray:
@@ -250,28 +201,6 @@ def get_mask_list(data, masks, is_gt):
     else:
         CACHE_DICTS['mask_list'][str(data) + str(masks) + str(is_gt)] = (bb_object, mask_list)
     return bb_object, mask_list
-
-
-def get_cat_instances_seg_lst(idx: int, data: PreprocessResponse, cat: str) -> List[np.ma.array]:
-    img = input_image(idx, data)
-    if cat == "tote":
-        masks = get_tote_instances_masks(idx, data)
-    elif cat == "object":
-        masks = get_object_instances_masks(idx, data)
-    else:
-        print('Error category not supported')
-        return None
-    if masks is None:
-        return None
-    if masks[0, ...].shape != IMAGE_SIZE:
-        masks = tf.image.resize(masks[..., None], IMAGE_SIZE, tf.image.ResizeMethod.NEAREST_NEIGHBOR)[..., 0]
-        masks = masks.numpy()
-    instances = []
-    for mask in masks:
-        mask = np.broadcast_to(mask[..., np.newaxis], img.shape)
-        masked_arr = np.ma.masked_array(img, mask)
-        instances.append(masked_arr)
-    return instances
 
 
 def remove_label_from_bbs(bbs_object_array, removal_label, add_to_label):
